# openapi/openapi-v1.yaml
openapi: 3.1.0
info:
  title: CorteCero API
  version: 1.0.0
  description: >
    API MVP para control de cut-off, planes diarios por zona, pedidos tardíos,
    excepciones y auditoría.
servers:
  - url: https://api.cortecero.local
...
security:
  - bearerAuth: []
paths:
...
  /plans:
    get:
      tags: [Plans]
      summary: Listado de planes por fecha/zona
      parameters:
        - in: query
          name: service_date
          schema:
            type: string
            format: date
        - in: query
          name: zone_id
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [open, locked, dispatched]
      responses:
        "200":
          description: Lista de planes
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Plan'
                  total:
                    type: integer

    post:
      tags: [Plans]
      summary: Crear plan para fecha y zona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_date, zone_id]
              properties:
                service_date:
                  type: string
                  format: date
                zone_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Plan creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
  /plans/{plan_id}:
    get:
      tags: [Plans]
      summary: Detalle de plan
      parameters:
        - in: path
          name: plan_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Plan encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
  /plans/{plan_id}/lock:
    post:
      tags: [Plans]
      summary: Bloquear un plan
      description: >
        Cambia el plan a locked. A partir de ese momento solo admite inclusión
        por excepción aprobada.
      parameters:
        - in: path
          name: plan_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Plan bloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
  /plans/{plan_id}/orders:
    post:
      tags: [Plans]
      summary: Incluir pedido en plan
      description: >
        Si el plan está open, admite pedidos ready_for_planning.
        Si el plan está locked, solo admite pedidos con excepción approved.
      parameters:
        - in: path
          name: plan_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id]
              properties:
                order_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Pedido incluido en plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanOrder'
        "409":
          description: Regla funcional impide inclusión
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Plan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        service_date:
          type: string
          format: date
        zone_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [open, locked, dispatched]
        version:
          type: integer
        locked_at:
          type: string
          format: date-time
          nullable: true
        locked_by:
          type: string
          format: uuid
          nullable: true
        orders:
          type: array
          items:
            $ref: '#/components/schemas/PlanOrder'
    PlanOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        inclusion_type:
          type: string
          enum: [normal, exception]
        added_at:
          type: string
          format: date-time
        added_by:
          type: string
          format: uuid
...